# chore: always inline all unexported native functions