--- a/internal/native/avx/native_subr_amd64.go
+++ b/internal/native/avx/native_subr_amd64.go
@@ -9,53 +9,53 @@ package avx
 func __native_entry__() uintptr
 
 var (
-    _subr__f32toa             = __native_entry__() + 29744
-    _subr__f64toa             = __native_entry__() + 496
-    _subr__get_by_path        = __native_entry__() + 27424
-    _subr__html_escape        = __native_entry__() + 9968
-    _subr__i64toa             = __native_entry__() + 4112
+    _subr__f32toa             = __native_entry__() + 51488
+    _subr__f64toa             = __native_entry__() + 448
+    _subr__get_by_path        = __native_entry__() + 44080
+    _subr__html_escape        = __native_entry__() + 10720
+    _subr__i64toa             = __native_entry__() + 5168
     _subr__lspace             = __native_entry__() + 80
-    _subr__quote              = __native_entry__() + 5792
-    _subr__skip_array         = __native_entry__() + 20576
-    _subr__skip_number        = __native_entry__() + 23920
-    _subr__skip_object        = __native_entry__() + 22496
-    _subr__skip_one           = __native_entry__() + 24080
-    _subr__skip_one_fast      = __native_entry__() + 24320
-    _subr__u64toa             = __native_entry__() + 4384
-    _subr__unquote            = __native_entry__() + 7488
-    _subr__validate_one       = __native_entry__() + 24144
-    _subr__validate_utf8      = __native_entry__() + 28464
-    _subr__validate_utf8_fast = __native_entry__() + 29136
-    _subr__value              = __native_entry__() + 14672
-    _subr__vnumber            = __native_entry__() + 18320
-    _subr__vsigned            = __native_entry__() + 19856
-    _subr__vstring            = __native_entry__() + 16864
-    _subr__vunsigned          = __native_entry__() + 20208
+    _subr__quote              = __native_entry__() + 6560
+    _subr__skip_array         = __native_entry__() + 24320
+    _subr__skip_number        = __native_entry__() + 34080
+    _subr__skip_object        = __native_entry__() + 26704
+    _subr__skip_one           = __native_entry__() + 35408
+    _subr__skip_one_fast      = __native_entry__() + 40272
+    _subr__u64toa             = __native_entry__() + 5280
+    _subr__unquote            = __native_entry__() + 8352
+    _subr__validate_one       = __native_entry__() + 37792
+    _subr__validate_utf8      = __native_entry__() + 50224
+    _subr__validate_utf8_fast = __native_entry__() + 50896
+    _subr__value              = __native_entry__() + 12000
+    _subr__vnumber            = __native_entry__() + 16096
+    _subr__vsigned            = __native_entry__() + 23536
+    _subr__vstring            = __native_entry__() + 13904
+    _subr__vunsigned          = __native_entry__() + 23904
 )
 
 const (
-    _stack__f32toa = 56
-    _stack__f64toa = 80
-    _stack__get_by_path = 312
+    _stack__f32toa = 48
+    _stack__f64toa = 56
+    _stack__get_by_path = 320
     _stack__html_escape = 64
     _stack__i64toa = 16
     _stack__lspace = 8
-    _stack__quote = 80
-    _stack__skip_array = 128
-    _stack__skip_number = 72
-    _stack__skip_object = 128
-    _stack__skip_one = 128
-    _stack__skip_one_fast = 208
+    _stack__quote = 56
+    _stack__skip_array = 120
+    _stack__skip_number = 56
+    _stack__skip_object = 120
+    _stack__skip_one = 120
+    _stack__skip_one_fast = 200
     _stack__u64toa = 8
-    _stack__unquote = 128
-    _stack__validate_one = 128
+    _stack__unquote = 88
+    _stack__validate_one = 104
     _stack__validate_utf8 = 48
     _stack__validate_utf8_fast = 24
-    _stack__value = 368
-    _stack__vnumber = 280
+    _stack__value = 184
+    _stack__vnumber = 128
     _stack__vsigned = 16
-    _stack__vstring = 128
-    _stack__vunsigned = 24
+    _stack__vstring = 104
+    _stack__vunsigned = 16
 )
 
 var (