… internal/native/avx/native_amd64.s
… internal/native/avx2/native_amd64.s
… internal/native/sse/native_amd64.s
… native/parsing.c
… native/scanning.c